# Improve perf and reduce allocs in `bsoncore.Document.LookupErr`.