@divjotarora 2020-04-17
Hi @bartle-stripe,

Thank you for the PR! I've authorized the CI tasks, but I don't anticipate any issues there. My understanding is that this saves allocations because the `string()` call you added in `LookupErr` doesn't store the result in a variable and therefore does not trigger a copy of the key bytes into an internal read-only string? Can you confirm?