@bartle-stripe 2020-04-18
Yes, go is smart enough to avoid an alloc here (and this is confirmed by the change in benchmark allocs).