@bartle-stripe 2020-04-18
Specifically, I think go's escape analysis is what prevents it from needing to be allocated; If you run `go build -gcflags=-m=1 x/bsonx/bsoncore/*.go` you can see:
```
x/bsonx/bsoncore/document.go:185:12: string(([]byte)(~r0)) does not escape
```